chore(cli parsing)!: cleanup cli parsing
Rass the CLI args from node over to java AS-IS. Removed all the yargs
hackery.

Root cause was that we were passing arrays as strings, and that required
all sorts of crazy escaping. `spawn` accepts array args, so we just
use that instead.

breaking change: removed kafka and graphql support